+++ RobotCasserole2016/src/org/usfirst/frc/team1736/robot/CasserolePID.java
@@ -91,7 +91,7 @@ public abstract class CasserolePID {
 		//Will start calling the periodic update function at an interval of pidSamplePeriod_ms,
 		//asynchronously from any other code.
 		//Java magic here, don't touch!
-        timerThread.scheduleAtFixedRate(new PIDTask(this), 0L, (long) (pidSamplePeriod_ms));
+        timerThread.scheduleAtFixedRate(new PIDTask(this), 0L, (pidSamplePeriod_ms));
 	}
 	
 	public void stop(){
